AWS Error: ./app/NewsletterPopup.tsx:30:14
2380

Type error: Type '{ backgroundColor: string; maxWidth: string; borderRadius: string; border: string; width: string; boxSizing: string; }' is not assignable to type 'Properties<string | number, string & {}>'.

2381

Types of property 'boxSizing' are incompatible.

Altered News Letter code attempt fix
