+++ app/page.tsx
@@ -12,7 +12,7 @@ import Link from "next/link";
 import CustomNavbar from "./CustomNavbar";
 import HPheader from "./HPheader";
 import ImageCarousel from "./ImageCarousel";
-import NewsletterPopup from "./NewsletterPopup";
+import NewsletterPopup from "./news-letter/NewsletterPopup";
 import ContactInfoBox from "./components/ContactInfoBox";
 import ContactUs from "./components/ContactUs";
 Amplify.configure(outputs);
